Update dashboard titles
