+++ pyproject.toml
@@ -18,6 +18,7 @@ matplotlib = "^3.10.1"
 matsim-tools = "^0.0.19"
 osmnx = "^2.0.1"
 pandas = "^2.2.3"
+python-slugify = "^8.0.4"
 pyyaml = "^6.0.2"
 rasterio = "^1.4.3"
 scikit-learn = "^1.6.1"
