+++ src/matsim_io/scripts.py
@@ -6,13 +6,13 @@ import yaml
 from matsim_io import MATSIM_DATA_DIR
 
 
-def move_dashboards(output_dirs: list[str]) -> None:
+def move_dashboards(output_dirs: list[tuple[str, str]]) -> None:
     """
     Move every dashboard-2.yaml file in the given directories up to the MATSim data directory.
-    :param output_dirs: List of subdirectories in the MATSim data directory.
+    :param output_dirs: List of tuples of MATSim subdirectories and SimWrapper titles.
     """
 
-    for i, output_dir in enumerate(output_dirs, start=1):
+    for i, (output_dir, title) in enumerate(output_dirs, start=1):
         output_path = MATSIM_DATA_DIR / output_dir
         if not output_path.is_dir():
             logging.error(f"Directory {output_path} does not exist.")
@@ -26,6 +26,7 @@ def move_dashboards(output_dirs: list[str]) -> None:
         dest_path = MATSIM_DATA_DIR / f"dashboard-{i}.yaml"
         dashboard_path.rename(dest_path)
         _rewrite_dashboard_dataset_paths(dest_path, output_dir)
+        _update_dashboard_title(dest_path, title)
 
 
 def _rewrite_dashboard_dataset_paths(dashboard: Path, output_dir: str) -> None:
